Refactor: switching dependency injection to constructor injection

+++ src/main/java/api/addressbook/controller/AddressController.java
@@ -1,8 +1,10 @@
 package api.addressbook.controller;
 
 import api.addressbook.entity.AddressEntity;
+import api.addressbook.mapper.AddressMapper;
 import api.addressbook.model.Address;
 import api.addressbook.repository.AddressRepository;
+import api.addressbook.repository.PersonRepository;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,27 +21,31 @@ import java.util.stream.Collectors;
 @Controller
 @RequestMapping("/address")
 public class AddressController extends AbstractController {
+    // should it be final?
 
-//    private final AddressRepository addressRepository;
-//
-//    public AddressController(AddressRepository addressRepository) {
-//        this.addressRepository = addressRepository;
-//    }
-//
+    private final AddressRepository addressRepository;
+
+    private final PersonRepository personRepository;
+
+    private final AddressMapper addressMapper;
 
     @Autowired
-    protected AddressRepository addressRepository;
+    public AddressController(AddressRepository addressRepository, PersonRepository personRepository, AddressMapper addressMapper) {
+        this.addressRepository = addressRepository;
+        this.personRepository = personRepository;
+        this.addressMapper = addressMapper;
+    }
 
     @RequestMapping("")
     @GetMapping(value = "/url", produces = "application/json")
-    public ResponseEntity<List<Address>> getAllPerson() {
-        logger.info("call for all address {}", addressRepository.findAll());
+    public ResponseEntity<List<Address>> getAllAddress() {
+        log.info("call for all address {}", addressRepository.findAll());
         return ResponseEntity.status(HttpStatus.FOUND).body(addressRepository.findAll().stream().map(addressMapper::toDomain).collect(Collectors.toList()));
     }
 
     @RequestMapping("/{id}")
     @GetMapping(value = "/url", produces = "application/json")
-    public ResponseEntity<Optional<Address>> getPersonById(@PathVariable("id") int id) {
+    public ResponseEntity<Optional<Address>> getAddressById(@PathVariable("id") int id) {
         if (addressRepository.findById(id).isPresent()) {
             return ResponseEntity.status(HttpStatus.FOUND).body(addressRepository.findById(id).map(addressMapper::toDomain));
         }
@@ -52,8 +58,8 @@ public class AddressController extends AbstractController {
     public ResponseEntity<Address> createAddress(@RequestBody @NonNull AddressEntity body) {
         if (!body.getStreetNumber().isBlank()) {
             Address address = addressMapper.toDomain(addressRepository.save(body));
-            personRepository.saveAll(body.getPerson());
-            logger.info("An address was added: {}", address);
+//            personRepository.saveAll(body.getPersonAddressEntity().);
+            log.info("An address was added: {}", address);
             return ResponseEntity.status(HttpStatus.CREATED).body(address);
         }
         return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
@@ -63,9 +69,9 @@ public class AddressController extends AbstractController {
     public ResponseEntity<Address> updateAddress(@PathVariable("id") Integer id, @RequestBody @NonNull AddressEntity body) {
         Address address = addressRepository.findById(id).map(ad -> {
             ad.setCountry(body.getCountry());
-            ad.setPrivate(body.isPrivate());
+            ad.setIsPrivate(body.getIsPrivate());
             ad.setLocality(body.getLocality());
-            ad.setPerson(body.getPerson());
+//            ad.setPerson(body.getPerson());
             ad.setStreetName(body.getStreetName());
             ad.setBoxNumber(body.getBoxNumber());
             ad.setStreetNumber(body.getStreetNumber());
@@ -73,13 +79,14 @@ public class AddressController extends AbstractController {
         }).map(addressMapper::toDomain).orElseGet(() -> {
             return addressMapper.toDomain(addressRepository.save(body));
         });
-        logger.info("An address has been updated: {}", address);
+        log.info("An address has been updated: {}", address);
         return ResponseEntity.status(HttpStatus.ACCEPTED).body(address);
     }
 
     @DeleteMapping(value = "/{id}")
     public ResponseEntity<Integer> deleteAllAddress(@PathVariable("id") Integer id) {
         if (addressRepository.existsById(id)) {
+            addressRepository.deleteById(id);
             return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
         }
         return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
@@ -89,6 +96,9 @@ public class AddressController extends AbstractController {
 
     @DeleteMapping("/deleteAll")
     public ResponseEntity<Integer> deleteAddressPerId() {
+        if (addressRepository.count() == 0) {
+            return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+        }
         addressRepository.deleteAll();
         return ResponseEntity.status(HttpStatus.ACCEPTED).build();
     }
